Improve vite bundling
Split all the big js files into many different modules imported at runtime
